Reorganized API around endpoints instead of class types
