+++ brambling/api/v1/urls.py
@@ -1,22 +1,20 @@
 from django.conf.urls import url, include
 from rest_framework.routers import DefaultRouter
 
-from brambling.api.v1.views import (
-    HousingCategoryViewSet,
-    EnvironmentalFactorViewSet,
-    DanceStyleViewSet,
-    OrganizationViewSet,
-    EventViewSet,
-    AttendeeViewSet,
-    EventHousingViewSet,
-    OrderViewSet,
-    OrderSearchViewSet,
-    BoughtItemViewSet,
-    ItemViewSet,
-    ItemImageViewSet,
-    ItemOptionViewSet,
-    OrderDiscountViewSet,
-)
+from brambling.api.v1.endpoints.attendee import AttendeeViewSet
+from brambling.api.v1.endpoints.boughtitem import BoughtItemViewSet
+from brambling.api.v1.endpoints.dancestyle import DanceStyleViewSet
+from brambling.api.v1.endpoints.environmentalfactor import EnvironmentalFactorViewSet
+from brambling.api.v1.endpoints.event import EventViewSet
+from brambling.api.v1.endpoints.eventhousing import EventHousingViewSet
+from brambling.api.v1.endpoints.housingcategory import HousingCategoryViewSet
+from brambling.api.v1.endpoints.item import ItemViewSet
+from brambling.api.v1.endpoints.itemimage import ItemImageViewSet
+from brambling.api.v1.endpoints.itemoption import ItemOptionViewSet
+from brambling.api.v1.endpoints.order import OrderViewSet
+from brambling.api.v1.endpoints.ordersearch import OrderSearchViewSet
+from brambling.api.v1.endpoints.orderdiscount import OrderDiscountViewSet
+from brambling.api.v1.endpoints.organization import OrganizationViewSet
 
 
 router = DefaultRouter()
